Route automation marker tags through the client tag channel
Every automation send was failing client-side: the marker tag rode the
outgoing media channel, whose builder rejects any non-imeta prefix —
so Run now toasted an ask (now an error, after the surfacing fix) but
no prompt ever reached the relay, the agent never started, and the
activation card sat on "Starting…" forever. The e2e mock accepted the
invalid tag, which is how the route stayed green while broken live.

The marker now rides the existing whitelisted ["client", ...] channel:
send_channel_message accepts client_tags and threads them to
build_message_with_client_tags; splitOutgoingTags gets a clientTags
bucket (client tags force the REST path so the tag-validating builder
runs); and the mock now enforces the imeta-only media gate and echoes
client tags, closing the test blind spot. The event→model converters
move to message_converters.rs to hold the size ceiling.
