+++ desktop/src-tauri/src/commands/messages.rs
@@ -1,6 +1,10 @@
 use nostr::{Event, EventId, Keys, PublicKey};
 use tauri::{AppHandle, State};
 
+use super::message_converters::{
+    feed_item_from_event, forum_message_from_event, forum_reply_from_event,
+};
+
 use crate::{
     app_state::AppState,
     events,
@@ -8,7 +12,7 @@ use crate::{
     models::{
         FeedItemInfo, FeedMeta, FeedResponse, FeedSections, ForumMessageInfo, ForumPostsResponse,
         ForumThreadReplyInfo, ForumThreadResponse, SearchResponse, SendChannelMessageResponse,
-        ThreadRepliesResponse, ThreadSummary,
+        ThreadRepliesResponse,
     },
     nostr_convert,
     relay::{query_relay, submit_event, submit_event_with_keys},
@@ -484,6 +488,7 @@ pub async fn send_channel_message(
     mention_tags: Option<Vec<Vec<String>>>,
     mention_pubkeys: Option<Vec<String>>,
     kind: Option<u32>,
+    client_tags: Option<Vec<Vec<String>>>,
     state: State<'_, AppState>,
 ) -> Result<SendChannelMessageResponse, String> {
     let channel_uuid = uuid::Uuid::parse_str(&channel_id)
@@ -493,6 +498,7 @@ pub async fn send_channel_message(
     let media = media_tags.unwrap_or_default();
     let emoji = emoji_tags.unwrap_or_default();
     let mention_refs_only = mention_tags.unwrap_or_default();
+    let client = client_tags.unwrap_or_default();
     let kind_num = kind.unwrap_or(buzz_core_pkg::kind::KIND_STREAM_MESSAGE);
 
     let mut resolved_root: Option<String> = None;
@@ -529,7 +535,7 @@ pub async fn send_channel_message(
                 }
                 None => None,
             };
-            events::build_message(
+            events::build_message_with_client_tags(
                 channel_uuid,
                 content.trim(),
                 thread_ref.as_ref(),
@@ -537,6 +543,7 @@ pub async fn send_channel_message(
                 &media,
                 &emoji,
                 &mention_refs_only,
+                &client,
             )?
         }
     };
@@ -986,96 +993,3 @@ pub async fn delete_message(
 }
 
 // ── Local helpers ───────────────────────────────────────────────────────────
-
-fn channel_id_from_tags(ev: &nostr::Event) -> Option<String> {
-    ev.tags.iter().find_map(|t| {
-        let s = t.as_slice();
-        if s.len() >= 2 && s[0] == "h" {
-            Some(s[1].clone())
-        } else {
-            None
-        }
-    })
-}
-
-fn tags_to_vec(ev: &nostr::Event) -> Vec<Vec<String>> {
-    ev.tags.iter().map(|t| t.as_slice().to_vec()).collect()
-}
-
-fn feed_item_from_event(ev: &nostr::Event, category: &str) -> FeedItemInfo {
-    let channel_id = channel_id_from_tags(ev);
-    FeedItemInfo {
-        id: ev.id.to_hex(),
-        kind: ev.kind.as_u16() as u32,
-        pubkey: ev.pubkey.to_hex(),
-        content: ev.content.clone(),
-        created_at: ev.created_at.as_secs(),
-        channel_id,
-        channel_name: String::new(),
-        channel_type: None,
-        tags: tags_to_vec(ev),
-        category: category.to_string(),
-    }
-}
-
-fn forum_message_from_event(ev: &nostr::Event, channel_id: &str) -> ForumMessageInfo {
-    ForumMessageInfo {
-        event_id: ev.id.to_hex(),
-        pubkey: ev.pubkey.to_hex(),
-        content: ev.content.clone(),
-        kind: ev.kind.as_u16() as u32,
-        created_at: ev.created_at.as_secs() as i64,
-        channel_id: channel_id.to_string(),
-        tags: tags_to_vec(ev),
-        thread_summary: Some(ThreadSummary {
-            reply_count: 0,
-            descendant_count: 0,
-            last_reply_at: None,
-            participants: Vec::new(),
-        }),
-        reactions: serde_json::Value::Null,
-    }
-}
-
-fn forum_reply_from_event(
-    ev: &nostr::Event,
-    channel_id: &str,
-    root_event_id: &str,
-) -> ForumThreadReplyInfo {
-    // Walk e-tags for NIP-10 parent/root markers.
-    let (mut parent_id, mut explicit_root) = (None, None);
-    for t in ev.tags.iter() {
-        let s = t.as_slice();
-        if s.len() >= 2 && s[0] == "e" {
-            match s.get(3).map(|x| x.as_str()) {
-                Some("root") => explicit_root = Some(s[1].clone()),
-                Some("reply") => parent_id = Some(s[1].clone()),
-                _ => {
-                    if parent_id.is_none() {
-                        parent_id = Some(s[1].clone());
-                    }
-                }
-            }
-        }
-    }
-    let parent = parent_id
-        .clone()
-        .unwrap_or_else(|| root_event_id.to_string());
-    let root = explicit_root.unwrap_or_else(|| root_event_id.to_string());
-    let depth = if parent == root { 1 } else { 2 };
-
-    ForumThreadReplyInfo {
-        event_id: ev.id.to_hex(),
-        pubkey: ev.pubkey.to_hex(),
-        content: ev.content.clone(),
-        kind: ev.kind.as_u16() as u32,
-        created_at: ev.created_at.as_secs() as i64,
-        channel_id: channel_id.to_string(),
-        tags: tags_to_vec(ev),
-        parent_event_id: Some(parent),
-        root_event_id: Some(root),
-        depth,
-        broadcast: false,
-        reactions: serde_json::Value::Null,
-    }
-}
