+++ desktop/src/features/messages/lib/imetaMediaMarkdown.ts
@@ -310,18 +310,25 @@ export function splitOutgoingTags(tags: string[][] | undefined): {
   mediaTags: string[][];
   emojiTags: string[][];
   mentionTags: string[][];
+  clientTags: string[][];
 } {
   const mediaTags: string[][] = [];
   const emojiTags: string[][] = [];
   const mentionTags: string[][] = [];
+  const clientTags: string[][] = [];
   for (const tag of tags ?? []) {
     if (tag[0] === "emoji") {
       emojiTags.push(tag);
     } else if (tag[0] === "mention") {
       mentionTags.push(tag);
+    } else if (tag[0] === "client") {
+      // Whitelisted client marker tags (e.g. automation prompts) — these
+      // must NOT ride the imeta-only media channel, whose builder rejects
+      // any non-imeta prefix and fails the whole send.
+      clientTags.push(tag);
     } else {
       mediaTags.push(tag);
     }
   }
-  return { mediaTags, emojiTags, mentionTags };
+  return { mediaTags, emojiTags, mentionTags, clientTags };
 }
